test: address CodeRabbit review - shared locale helper, comment cleanup
Extract the duplicated en.json dotted-key resolver + i18n mock setup
from launchSettingsFields.test.ts and registerSettingsHandlers.test.ts
into src/main/lib/localeTestHelper.ts so the copies cannot diverge.
Rewrite comments that narrated the setting's former global placement to
describe the current per-install behavior instead.

Amp-Thread-ID: https://ampcode.com/threads/T-019f90ea-5c63-76c9-8608-0043b9fdbc94
